moved config providers to separate assemblies and loaded via reflection
