+++ NETFundamentals.sln
@@ -24,7 +24,13 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "FileSystemVisitorTests", "A
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Reflection", "Reflection", "{C197E5EB-64E7-4B9D-BC19-7793DBA3CC96}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CustomConfigurationDemo", "Reflection\CustomConfigurationDemo\CustomConfigurationDemo.csproj", "{85F239AD-F447-4BA6-994D-9556987D594F}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CustomConfigurationDemo", "Reflection\CustomConfigurationDemo\CustomConfigurationDemo.csproj", "{85F239AD-F447-4BA6-994D-9556987D594F}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FileConfigurationProviderProject", "Reflection\FileConfigurationProviderProject\FileConfigurationProviderProject.csproj", "{88E9E054-7A98-4B2E-ADB3-33BF265EB25E}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ConfigurationManagerConfigurationProviderProject", "Reflection\ConfigurationManagerConfigurationProviderProject\ConfigurationManagerConfigurationProviderProject.csproj", "{C287A7B0-B3BA-4F57-9390-3A61E3FAD26C}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CommonProvidersProject", "Reflection\CommonProvidersProject\CommonProvidersProject.csproj", "{E04077FE-AA67-4992-8AC0-B965B2196C22}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -64,6 +70,18 @@ Global
 		{85F239AD-F447-4BA6-994D-9556987D594F}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{85F239AD-F447-4BA6-994D-9556987D594F}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{85F239AD-F447-4BA6-994D-9556987D594F}.Release|Any CPU.Build.0 = Release|Any CPU
+		{88E9E054-7A98-4B2E-ADB3-33BF265EB25E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{88E9E054-7A98-4B2E-ADB3-33BF265EB25E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{88E9E054-7A98-4B2E-ADB3-33BF265EB25E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{88E9E054-7A98-4B2E-ADB3-33BF265EB25E}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C287A7B0-B3BA-4F57-9390-3A61E3FAD26C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C287A7B0-B3BA-4F57-9390-3A61E3FAD26C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C287A7B0-B3BA-4F57-9390-3A61E3FAD26C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C287A7B0-B3BA-4F57-9390-3A61E3FAD26C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{E04077FE-AA67-4992-8AC0-B965B2196C22}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E04077FE-AA67-4992-8AC0-B965B2196C22}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E04077FE-AA67-4992-8AC0-B965B2196C22}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E04077FE-AA67-4992-8AC0-B965B2196C22}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -77,6 +95,9 @@ Global
 		{601389D2-FF96-47F8-8AE3-8E6577C276FD} = {C5AA7F2E-6733-4B87-92A4-E74F2C7CA3B4}
 		{16B32515-FE27-47A3-A96C-5A3A5AE167A0} = {C5AA7F2E-6733-4B87-92A4-E74F2C7CA3B4}
 		{85F239AD-F447-4BA6-994D-9556987D594F} = {C197E5EB-64E7-4B9D-BC19-7793DBA3CC96}
+		{88E9E054-7A98-4B2E-ADB3-33BF265EB25E} = {C197E5EB-64E7-4B9D-BC19-7793DBA3CC96}
+		{C287A7B0-B3BA-4F57-9390-3A61E3FAD26C} = {C197E5EB-64E7-4B9D-BC19-7793DBA3CC96}
+		{E04077FE-AA67-4992-8AC0-B965B2196C22} = {C197E5EB-64E7-4B9D-BC19-7793DBA3CC96}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {E67B29C8-325A-4DBD-A85A-56B7332CE3EB}
