+++ Reflection/CommonProvidersProject/IConfigurationProvider.cs
@@ -1,0 +1,8 @@
+﻿namespace CommonProvidersProject;
+
+public interface IConfigurationProvider
+{
+    void SetValue(string key, object value);
+
+    object GetValue(string key, Type valueType);
+}
