+++ Reflection/CustomConfigurationDemo/Attributes/BaseConfigurationItemAttribute.cs
@@ -1,9 +1,4 @@
-﻿using CustomConfigurationDemo.Providers;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿using CommonProvidersProject;
 
 namespace CustomConfigurationDemo.Attributes;
