+++ Reflection/FileConfigurationProviderProject/FileConfigurationProviderProject.csproj
@@ -1,0 +1,13 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net6.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\CommonProvidersProject\CommonProvidersProject.csproj" />
+  </ItemGroup>
+
+</Project>
